JAVA-13-Solution/src/com/zeuschiu/generation/solution/service/StudentService.java
package com.zeuschiu.generation.solution.service;

import com.zeuschiu.generation.solution.model.Course;
import com.zeuschiu.generation.solution.model.Student;

import java.util.HashMap;
import java.util.Map;

public class StudentService
{
    private final Map<String, Student> students = new HashMap<>();

    public void subscribeStudent( Student student )
    {
        students.put( student.getId(), student );
    }

    public Student findStudent( String studentId )
    {
        if ( students.containsKey( studentId ) )
        {
            return students.get( studentId );
        }
        return null;
    }

    public boolean isSubscribed( String studentId )
    {
        return students.containsKey( studentId );
    }

    public void showSummary()
    {
        for(Student student:students.values()){
            System.out.println(student.toString());
        }
    }

    public void enrollToCourse( String studentId, Course course )
    {
        if ( students.containsKey( studentId ) )
        {
            students.get( studentId ).enrollToCourse( course );
        }
    }


}
